requirements.txt
flask==3.0.3
gunicorn==22.0.0
jinja2==3.1.4
python-dotenv==1.0.1

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.11.4
joblib==1.4.2
matplotlib==3.8.4

SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
requests==2.32.3


